refactor(server): Add base types and update image helpers
This commit introduces new base types and includes updates to the image helper functions.

+++ server/src/utils/image.helpers.ts
@@ -1,5 +1,20 @@
 import path from "path";
 import fs from "fs/promises";
+import {UpdateProductImageDto} from "@src/types/dto/product/UpdateProductDto";
+import { ProductImage } from "../../generated/prisma";
+import * as process from "node:process";
+
+
+/**
+ * Split images into groups
+ */
+export function splitImages(images: UpdateProductImageDto[]) {
+    const existingImages = images.filter((img) => img.id);
+    const newImages = images.filter((img) => !img.id);
+    const existingImagesId = existingImages.map(img => img.id!);
+
+    return { existingImages, newImages, existingImagesId };
+}
 
 /**
  * Extract filename from full URL
@@ -14,16 +29,32 @@ export function extractFilenameFromUrl(url: string): string | null {
 }
 
 /**
- * Delete image file from local storage
+ * Find images to delete
  */
-export async function deleteImageFile(filename: string): Promise<boolean> {
-    try {
-        const filePath = path.join(process.cwd(), "uploads", "products", filename);
-        await fs.unlink(filePath);
-        return true;
-    } catch {
-        return false;
-    }
+export function findImagesToDelete(
+    existingProductImages: ProductImage[],
+    keepImageIds: string[]
+) {
+    return existingProductImages.filter((img) => !keepImageIds.includes(img.id));
+}
+
+/**
+ * Delete files from local storage
+ */
+export async function deleteImageFiles(images: ProductImage[]) {
+    await Promise.all(
+        images.map(async (img) =>{
+            try {
+                const filename = extractFilenameFromUrl(img.url);
+                if (filename) {
+                    const filePath = path.join(process.cwd(), "uploads", "products", filename);
+                    await fs.unlink(filePath);
+                }
+            } catch (error) {
+                console.error(`⚠️ Failed to delete file for image: ${img.url}`, error);
+            }
+        })
+    )
 }
 
 /**
